Reorganize ai package
