+++ project/game/ai/strategies/chinitsu.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from game.ai.first_version.strategies.honitsu import HonitsuStrategy
-from game.ai.first_version.strategies.main import BaseStrategy
+from game.ai.strategies.honitsu import HonitsuStrategy
+from game.ai.strategies.main import BaseStrategy
 from mahjong.tile import TilesConverter
 from mahjong.utils import (
     count_tiles_by_suits,
