+++ project/game/ai/tests/strategies/tests_formal_tempai.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 import unittest
 
-from game.ai.first_version.strategies.formal_tempai import FormalTempaiStrategy
-from game.ai.first_version.strategies.main import BaseStrategy
+from game.ai.strategies.formal_tempai import FormalTempaiStrategy
+from game.ai.strategies.main import BaseStrategy
 from game.table import Table
 from mahjong.meld import Meld
 from mahjong.tests_mixin import TestMixin
