+++ project/game/ai/tests/tests_ai.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import unittest
 
-from game.ai.first_version.strategies.main import BaseStrategy
+from game.ai.strategies.main import BaseStrategy
 from game.table import Table
 from mahjong.meld import Meld
 from mahjong.tests_mixin import TestMixin
